Restructure Docker build into dedicated subproject
- Move Dockerfile, DOCKER.md to docker/ directory
- Extract build script to docker/build-xdk.sh (clean, no branchy if/else)
- Create docker/build.gradle.kts with all Docker tasks
- Forward Docker tasks from main build.gradle.kts to docker subproject
- Use Bellsoft JDK consistently (remove Temurin)
- Use commit SHA for VERSION build arg (not branch name)
- Build for native platform (no more cross-platform issues)
- Fix Alpine package manager (apk not apt-get)
